make verify command accept stdin input
This allow to pipe data in and out of the verify command,
ensuring that errors are outputted on stderr.

For example:

	prmd verify < schema.json 2> errors

This also improve error formatting, by displaying file path
along with errors, allowing to filter errors per file easily.

For example:

       prmd verify schema/ |& grep account

This will also allow to use Prmd.verify() outside of the cli.