change the combine command to remove "id" from per schema
See also the "7. URI resolution scopes and dereferencing" chapter in
JSON Schema core definitions draft v4. The "id" attribute is to define
its scope. This "scope" is used for resolving references.

prmd combine applys "id" to per combined schema, but included "$ref"
values point to paths which are based in its root schema. The combined
schema is understood that it expects a single scope, so "id" attributes
should be removed on combining.