Link to resources from attributes list
When a resource's property refers to another resource, create links in
the first resource's attributes list to the second resource.

When a property is a nullable object consisting of properties from other
resources, and if all the properties refer to the same resource, the
nullable object's entry will also link to that resource, even if the
nullable object wasn't a direct reference.